make parse_router public in rfs for other tool
